+++ projects/net.splitcells.dem.api/src/main/java/net/splitcells/dem/data/set/legacy/LegacySetFactory.java
@@ -1,7 +1,7 @@
 /* SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-or-later
  * SPDX-FileCopyrightText: Contributors To The `net.splitcells.*` Projects
  */
-package net.splitcells.dem.data.set;
+package net.splitcells.dem.data.set.legacy;
 
 import net.splitcells.dem.lang.annotations.JavaLegacy;