+++ projects/net.splitcells.dem.api/src/main/java/net/splitcells/dem/data/set/map/Map.java
@@ -30,7 +30,7 @@ import java.util.function.Supplier;
 
 import static net.splitcells.dem.data.Flows.flow;
 import static net.splitcells.dem.data.order.Ordering.EQUAL;
-import static net.splitcells.dem.data.set.SetLegacyWrapper.setLegacyWrapper;
+import static net.splitcells.dem.data.set.legacy.SetLegacyWrapper.setLegacyWrapper;
 import static net.splitcells.dem.data.set.Sets.setOfUniques;
 import static net.splitcells.dem.data.set.map.Maps.map;
 import static net.splitcells.dem.lang.tree.TreeI.tree;